temp: add int128_t() around numbers

+++ src/omnicore/mdex.cpp
@@ -204,7 +204,7 @@ static MatchReturnType x_Trade(CMPMetaDEx* const pnew)
 
             // First determine how many representable (indivisible) tokens Alice can
             // purchase from Bob, using Bob's unit price
-            rational_t rCouldBuy = pnew->getAmountRemaining() * rational_t(pold->getAmountForSale(), pold->getAmountDesired());
+            rational_t rCouldBuy = int128_t(pnew->getAmountRemaining()) * rational_t(int128_t(pold->getAmountForSale()), int128_t(pold->getAmountDesired()));
 
             // This implies rounding down, since rounding up is impossible, and would
             // require more tokens than Alice has
